#14 UX: When flipping a cover with the A button, only B button can go back, A should be a toggle
It was super confusing the first time, I had thought I was stuck. Because I hit A to flip it I expected A to flip it back. A or B should be usable here. Probably an easy fix.

So what I ended up doing was getting rid of the flipped cover alternate mode altogether as it's not very useful. It lets you zoom and stuff but the textures are high resolution enough to let you see any more info than not zooming.

So now you hover over the game cover, hit the A button to flip it, then it will flip in place and nothing else changes. Then just hover over it and hit A again to flip it back. I think that's much nicer UX.