Remove 20 second wait before Wiiflow boots from SD #7
BOOM! Did it the right way. So first it checks if the app folder is on SD, then it checks if the config file exists, then it checks if data_on_usb is true, then it checks each section to see if it's enabled and if it's partition is 1 or higher (i.e. on USB). If any of those are true, then it assumes we're using USB and mounts USB. If they're all false, we can safely assume we're using SD only and it skips trying to mount USB and waiting 20 seconds for a drive that doesn't exist.

So now, if you have an all SD card setup properly configured, it will automatically know to skip USB mounting and will launch as quickly as someone with a USB drive. Actually a bit faster since it can skip the mounting step.